fix plackett-luce to properly handle tied ranks
